+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 All notable changes to the "dev-tools" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [v0.0.2]
+- Update icon for package and README
+- Add command convert for mac
 
-## [Unreleased]
-
+## [v0.0.1]
 - Initial release